pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-qna-agent"
version = "0.1.0"
description = "A retrieval-enabled agent for book Q&A using OpenAI Agents SDK with Qdrant vector database integration"
authors = [{name = "AI Engineer", email = "engineer@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "cohere>=5.0.0",
    "qdrant-client>=1.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "pytest>=7.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=0.950",
    "pylint>=2.14.0"
]

[project.scripts]
book-qna-agent = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]